collector: fix bug where a single silent logoff for an IP destroyed a session
This was fine if the session only had a single IP associated with
it, but for sessions with multiple IPs (e.g. v4 and v6) we were
erroneously removing the whole session when only one of the IPs
was subject to a silent logoff.

Now we just remove the IP that has switched owners from the session
and if that was the last remaining IP for the session, then we
can go ahead and destroy the whole session.