collector: fix memory access bug when tracking IP -> session mappings
This bug would often manifest when a session has been deleted. The
IP address for the session would be freed, but we were also using
that address as the key in our session map. This would lead to
errors when searching the map, as keys would now be invalid and
could have bogus links to other members of the map.