Add code to support X3 IPv4 PDUs
Modified and/or renamed a bunch of encoding job creation methods
to ensure a clean distinction between methods that take a
libtrace packet as input vs those that take a pointer to the
start of the intercepted "content".

Moved some of these methods to different source files that make
more sense for them to exist in.